Added custom Swiftgen templates to hide all styling behind protocols declated in `StylingProtocols`
