+++ Flickr Image Gallery/Flickr Image Gallery/Support Files/Base.lproj/Localizable.strings
@@ -7,4 +7,4 @@
 */
 
 
-"photolist.title" = "Photo Feed";
+"photo_list.title" = "Photo Feed";
